added final features for the site

+++ posts/views.py
@@ -1,16 +1,21 @@
 from urllib.parse import quote_plus
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
+from django.utils import timezone
+from django.db.models import Q
 from .models import Post
 from .forms import PostForm
 
 # Create your views here.
 def post_create(request):
+    if not request.user.is_staff or  not request.user.is_superuser:
+        raise Http404
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
+        instance.user = request.user
         instance.save()
         messages.success(request, 'Successfully Created!')
         return HttpResponseRedirect(instance.get_absolute_url())
@@ -22,6 +27,9 @@ def post_create(request):
 
 def post_detail(request, id):
     instance = get_object_or_404(Post, id=id)
+    if instance.draft or instance.publish > timezone.now().date():
+        if not request.user.is_staff or not request.user.is_superuser:
+            raise Http404
     share_string = quote_plus(instance.content)
     context = {
         'title': instance.title,
@@ -32,7 +40,19 @@ def post_detail(request, id):
 
 
 def post_list(request):
-    queryset_list = Post.objects.all()
+    today = timezone.now().date()
+    queryset_list = Post.objects.active()
+    if request.user.is_staff or request.user.is_superuser:
+        queryset_list = Post.objects.all()
+
+    query = request.GET.get('q')
+    if query:
+        queryset_list = queryset_list.filter(
+            Q(title__icontains=query) |
+            Q(content__icontains=query) |
+            Q(user__first_name__icontains=query) |
+            Q(user__last_name__icontains=query)
+        ).distinct()
     paginator = Paginator(queryset_list, 2)
     page_request_var = 'page'
     page = request.GET.get(page_request_var)
@@ -47,16 +67,20 @@ def post_list(request):
     context = {
         'object_list': queryset,
         'title': 'List',
-        'page_request_var': page_request_var
+        'page_request_var': page_request_var,
+        'today': today
     }
     return render(request, 'post_list.html', context)
 
 
 def post_update(request, id=None):
+    if not request.user.is_staff or  not request.user.is_superuser:
+        raise Http404
     instance = get_object_or_404(Post, id=id)
     form = PostForm(request.POST or None, request.FILES or None, instance=instance)
     if form.is_valid():
         instance = form.save(commit=False)
+        instance.user = request.user
         instance.save()
         messages.success(request, "<a href='#'>Item</a> Saved", extra_tags='html_safe')
         return HttpResponseRedirect(instance.get_absolute_url())
@@ -69,7 +93,10 @@ def post_update(request, id=None):
 
 
 def post_delete(request, id=None):
+    if not request.user.is_staff or  not request.user.is_superuser:
+        raise Http404
     instance = get_object_or_404(Post, id=id)
+    instance.user = request.user
     instance.delete()
     messages.success(request, 'Successfully Deleted')
     return redirect('posts:list')
